Bash Script to automate creating mysql tables
This file has to be executable, so run chmod+x on it in order to run it via bash. It takes one argument which is the table name you want to create. The user will then be promted for column names and definitions. The full query will be echoed out so you can decide if you want to run it or not. If you enter 'y' , the query will run, and the schema will be printed out for verification.